Revise matching of plumbing rules
The default rules are now handled like the ones defined in
g:acme_plumbing.

If the return value of a handler is a string it is run as a command and
the status value is used as the result, otherwise the return value is
directly used as the result.

This fixes matching 'file:pos' selections, which did not work because
the regex '\f+([:]...)?' failed in combination with &isfname set to
1-255. Changing '+' to non-greedy breaks normal right-clicking. The
proper solution is to have multiple rules, one for paths without
positions and one for paths with postions that are no longer optional.