Revise cmd prompt in acmegit
Use a new msg to overwrite the prompt after each click and show all the
added arguments in the prompt. Provide suggestions below the prompt.

This is much nicer if a list function provides lot of output because the
prompt is no longer hidden at the bottom of the buffer.