Improve guessing split area when drag-splitting
Before this the split happened along the window edge closest to the
mouse pointer. This did not work well for very wide or very tall
windows. The split is now opened in the area under the pointer that is
furthest away from the center of the window. This almost always works as
expected regardless of the shape of the window.