Simplify finding files
With this paths must be absolute or relative to the directory of the
current file; thus, opening files is done in the same way as running
commands.

CPP includes are no longer right-clickable, which wasn't very useful
anymore, because of special include directories (e.g. C++, cross-build).

Also gone is support for GNU make's -C directory messages, so that
right-clicking on relative paths in compiler messages may not work.