Revise matching ripgrep output again
This much simpler matching works reliable with my wrappers g and gw. It
will fail to match filenames in the output of rg(1) that look like a
match, i.e. begin with a number followed by a colon or minus sign. But
they will fail and not open the wrong file like older versions.

I prefer this over the previous more complex version which tested every
line as a filename because it works a 100% reliable with my wrappers and
does a good job on the default output of rg(1).