Improve branch suggestions in acmegit
Sort the branches by date, show only the branch names and filter out
branches pointing to the same commit as the one before them. Luckily,
git gives the local branches first and always groups remote branches by
the name of the remotes so that a lot of sync state is given by the fact
which branches are listed.