Overhaul moving windows
This merges the two ways of moving windows. A window is now only moved
by dragging its status bar with the right mouse button. If the dragged
win and the one over which the mouse button is released are in the same
column then the dragged win is moved up or down like it was done before
with the mouse wheel over a status bar. Otherwise the dragged win is
reopened as a split of the win under the mouse. Also the height of the
dragged win is not increased and it is not focused anymore.

With this we can get rid of all the scroll wheel mappings.