Overhaul win zooming
This makes it easy to quickly and temporarily rearrange wins.

Keep a small number of the current or important wins at the top of each
column and the occasionally needed ones minimized at the bottom. I
typically have a left column with the files I am editing and a right
column with ^agit, ^alsp, ^afx (or a couple of dir bufs) and a couple of
minimized guide files.

Right-clicking a status bar resizes the wins above it equally upwards.
But when a buf needs less lines the saved space is taken upwards too.
This means that the wins at the top can be much bigger then the lower
ones.

Right-clicking a minimized win at the bottom gives it some space and
possibly some other minimized wins above it too. Minimizing them again
is now easily done by dragging the status bar above them as far below as
possible and then right-clicking it. Without the right-click all the
space is taken up by the win in the middle whose status bar is dragged
and one has to instead drag the status bars of the wins above it one by
one to rearrange them in the way they were before.

It is best to accept the result of a right-click or rearrange the order
of the wins so that a better result can be achieved by it instead of
manually dragging a lot of status bars. It is also best to simply
right-click the status bar of a small +Errors win or scratch buf after
its job has produced some output to give it more space by taking it
equally from the ones above it.