Revise middle-click and right-click on statusline
Wins are now closed by middle-clicking the box on the left. This makes
it much less likely to accidentally close minimized guide wins when
trying to execute some cmd in them by middle-clicking it.
Middle-clicking anywhere else in the statusline kills a win's jobs.

Right-clicking the box on the left opens the win in a new column on the
left.