Limit right-click to plumbing
The user experience of also doing a text search if no plumbing rule
matched was too bad. If a right-click is meant to open something it is
not helpful to do a search and there was no way to do a search directly
in case the text was matched by a plumbing rule.