Revise scrolling in cleared scratch bufs
With this the cursor is kept on the first char of the last line in agit
and alsp bufs. But sending text to the client does not move the cursor
to this cell, which means that hints below the prompt get not scrolled.

This also makes it possible to see when a quick full screen update
happens because the cursor was moved.